cleanup old files and consolidate
